Clean config folder. Separate tasks into folders
