+++ config/favicon/faviconDescription.json
