+++ config/scripts/.babelrc
