+++ config/scripts/.eslintrc
