+++ config/styles/stylelint.config.js
